Separate viewholder from activity
