+++ app/src/main/java/co/omisego/omgshop/pages/history/TransactionHistoryActivity.kt
@@ -1,31 +1,24 @@
 package co.omisego.omgshop.pages.history
 
 import android.os.Bundle
-import android.support.v4.content.ContextCompat
 import android.support.v7.util.DiffUtil
 import android.support.v7.widget.DividerItemDecoration
 import android.support.v7.widget.LinearLayoutManager
 import android.support.v7.widget.RecyclerView
 import android.view.LayoutInflater
 import android.view.MenuItem
-import android.view.View
 import android.view.ViewGroup
 import co.omisego.omgshop.R
 import co.omisego.omgshop.base.BaseActivity
 import co.omisego.omgshop.extensions.logi
 import co.omisego.omgshop.pages.checkout.caller.TransactionHistoryCallerContract
-import co.omisego.omisego.model.pagination.Paginable
-import co.omisego.omisego.model.pagination.Paginable.Transaction.TransactionStatus
+import co.omisego.omgshop.pages.history.viewholder.LoadMoreViewCommand
+import co.omisego.omgshop.pages.history.viewholder.LoadMoreViewHolder
+import co.omisego.omgshop.pages.history.viewholder.TransactionHistoryViewHolder
 import co.omisego.omisego.model.transaction.list.Transaction
-import co.omisego.omisego.model.transaction.list.TransactionSource
 import kotlinx.android.synthetic.main.activity_transaction_history.*
-import kotlinx.android.synthetic.main.viewholder_load_more.view.*
-import kotlinx.android.synthetic.main.viewholder_transaction_record.view.*
-import java.math.RoundingMode
-import java.text.SimpleDateFormat
-import java.util.*
 
-class TransactionHistoryActivity : BaseActivity<TransactionHistoryContract.View, TransactionHistoryCallerContract.Caller, TransactionHistoryContract.Presenter>(), TransactionHistoryContract.View {
+class TransactionHistoryActivity : BaseActivity<TransactionHistoryContract.View, TransactionHistoryCallerContract.Caller, TransactionHistoryContract.Presenter>(), TransactionHistoryContract.View, LoadMoreViewCommand {
     override val mPresenter: TransactionHistoryContract.Presenter by lazy { TransactionHistoryPresenter() }
     private var myAddress: String = ""
     private val transactionListAdapter by lazy { TransactionHistoryAdapter() }
@@ -42,6 +35,10 @@ class TransactionHistoryActivity : BaseActivity<TransactionHistoryContract.View,
             transactionListAdapter.clear()
             loadTransactions()
         }
+    }
+
+    override fun loadMore() {
+        loadTransactions(currentPage)
     }
 
     private fun loadTransactions(page: Int = 1) {
@@ -96,17 +93,16 @@ class TransactionHistoryActivity : BaseActivity<TransactionHistoryContract.View,
             return when (viewType) {
                 typeItem -> {
                     val view = LayoutInflater.from(parent.context).inflate(R.layout.viewholder_transaction_record, parent, false)
-                    TransactionHistoryViewHolder(view)
+                    TransactionHistoryViewHolder(view, myAddress)
                 }
                 else -> {
                     val view = LayoutInflater.from(parent.context).inflate(R.layout.viewholder_load_more, parent, false)
-                    val loadMoreViewHolder = LoadMoreViewHolder(view)
+                    val loadMoreViewHolder = LoadMoreViewHolder(view, this@TransactionHistoryActivity)
                     loadingListener = loadMoreViewHolder
                     return loadMoreViewHolder
                 }
             }
         }
-
 
         override fun getItemViewType(position: Int): Int {
             return if (transactionList.size == position || transactionList.isEmpty()) typeLoadMore
@@ -142,96 +138,6 @@ class TransactionHistoryActivity : BaseActivity<TransactionHistoryContract.View,
             val diffResult = DiffUtil.calculateDiff(transactionHistoryDiffCallback)
             diffResult.dispatchUpdatesTo(this)
             this.transactionList.clear()
-        }
-
-        inner class TransactionHistoryViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
-            private val dateFormat by lazy { SimpleDateFormat("dd MMM yyyy HH:mm:ss", Locale.US) }
-
-            fun bindItem(record: Transaction) {
-                with(record) {
-                    val isSameAddress = isSameAddress(record.from)
-
-                    setTransactionDirection(isSameAddress)
-                    colorizedTransactionAmount(isSameAddress)
-                    record.setTransactionAddress(isSameAddress)
-                    record.status.colorizedTransactionStatus()
-                    record.from.formatTransactionAmount(isSameAddress)
-                    itemView.tvTransactionStatus.text = "- ${record.status.value.capitalize()}"
-                    itemView.tvTransactionDate.text = dateFormat.format(record.createdAt)
-                }
-            }
-
-            private fun Transaction.setTransactionAddress(sameAddress: Boolean) {
-                itemView.tvTransactionAddress.text = if (sameAddress) {
-                    this.to.address
-                } else {
-                    this.from.address
-                }
-            }
-
-            private fun colorizedTransactionAmount(sameAddress: Boolean) {
-                val color = if (sameAddress) {
-                    ContextCompat.getColor(itemView.context, R.color.colorRed)
-                } else {
-                    ContextCompat.getColor(itemView.context, R.color.colorGreen)
-                }
-
-                itemView.tvTransactionAmount.setTextColor(color)
-            }
-
-            private fun Paginable.Transaction.TransactionStatus.colorizedTransactionStatus() {
-                val color = when (this) {
-                    TransactionStatus.PENDING -> ContextCompat.getColor(itemView.context, R.color.colorYellow)
-                    TransactionStatus.CONFIRMED -> ContextCompat.getColor(itemView.context, R.color.colorGreen)
-                    TransactionStatus.FAILED, TransactionStatus.UNKNOWN -> ContextCompat.getColor(itemView.context, R.color.colorRed)
-                }
-
-                itemView.tvTransactionStatus.setTextColor(color)
-            }
-
-            private fun TransactionSource.formatTransactionAmount(sameAddress: Boolean) {
-                val amount = String.format("%.1f", this.amount.divide(this.token.subunitToUnit, RoundingMode.FLOOR))
-                val prefix = if (sameAddress) {
-                    "-"
-                } else {
-                    "+"
-                }
-                itemView.tvTransactionAmount.text = "$prefix $amount OMG"
-            }
-
-            private fun setTransactionDirection(sameAddress: Boolean) {
-                val direction = if (sameAddress) "To" else "From"
-                itemView.tvTransactionDirection.text = direction
-            }
-
-            private fun isSameAddress(source: TransactionSource) = myAddress == source.address
-        }
-
-        inner class LoadMoreViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView), ItemLoadingListener {
-            fun bindClick() {
-                itemView.tvLoadMore.text = getString(R.string.load_more, mPresenter.getPerPage())
-                itemView.setOnClickListener {
-                    setViewLoading(true)
-                    val request = mPresenter.createTransactionListParams(currentPage + 1)
-                    mPresenter.caller?.loadTransactionList(request)
-                }
-            }
-
-            override fun onFinished() {
-                setViewLoading(false)
-            }
-
-            private fun setViewLoading(loading: Boolean) {
-                if (loading) {
-                    itemView.tvLoadMore.visibility = View.INVISIBLE
-                    itemView.progressBar.visibility = View.VISIBLE
-                    itemView.isEnabled = false
-                } else {
-                    itemView.isEnabled = true
-                    itemView.tvLoadMore.visibility = View.VISIBLE
-                    itemView.progressBar.visibility = View.INVISIBLE
-                }
-            }
         }
     }
