+++ app/src/main/java/co/omisego/omgshop/pages/history/TransactionHistoryPresenter.kt
@@ -7,9 +7,7 @@ import co.omisego.omgshop.pages.checkout.caller.TransactionHistoryCallerContract
 import co.omisego.omgshop.pages.history.caller.TransactionHistoryCaller
 import co.omisego.omisego.model.APIError
 import co.omisego.omisego.model.OMGResponse
-import co.omisego.omisego.model.pagination.Paginable
 import co.omisego.omisego.model.pagination.PaginationList
-import co.omisego.omisego.model.pagination.SortDirection
 import co.omisego.omisego.model.transaction.list.Transaction
 import co.omisego.omisego.model.transaction.list.TransactionListParams
 
@@ -23,19 +21,15 @@ class TransactionHistoryPresenter : BasePresenterImpl<TransactionHistoryContract
     TransactionHistoryContract.Presenter,
     TransactionHistoryCallerContract.Handler {
 
-    private val perPage = 5
-
     override fun createTransactionListParams(page: Int): TransactionListParams {
         return TransactionListParams.create(
             page = page,
-            perPage = perPage,
-            sortBy = Paginable.Transaction.SortableFields.CREATED_AT,
-            sortDir = SortDirection.DESCENDING,
+            perPage = PaginationConfig.PER_PAGE,
+            sortBy = PaginationConfig.SORT_BY,
+            sortDir = PaginationConfig.SORT_DIR,
             searchTerm = null
         )
     }
-
-    override fun getPerPage() = perPage
 
     override fun loadCurrentAddress() {
         mView?.showCurrentAddress(Preference.loadWalletAddress())
