+++ app/src/main/java/co/omisego/omgshop/pages/history/viewholder/LoadMoreViewCommand.kt
@@ -1,0 +1,12 @@
+package co.omisego.omgshop.pages.history.viewholder
+
+
+/*
+ * OmiseGO
+ *
+ * Created by Phuchit Sirimongkolsathien on 22/6/2018 AD.
+ * Copyright © 2017-2018 OmiseGO. All rights reserved.
+ */
+interface LoadMoreViewCommand {
+    fun loadMore()
+}
